+++ test/testdata/workspaces/typescript-sample-app-lint-error/src/index.ts
@@ -1,6 +1,6 @@
 import http = require("http");
 
-const requestListener = (req: http.IncomingMessage, res: http.ServerResponse) => {
+const requestListener = (req: any, res: http.ServerResponse) => {
     res.writeHead(200);
     res.end('Hello, World!');
 }