+++ CHANGELOG.md
@@ -20,6 +20,7 @@ listed in the changelog.
 - Skipping Tests in TypeScript build task if test artifacts are present already ([#238](https://github.com/opendevstack/ods-pipeline/issues/238))
 - Provided make target for ShellCheck and added ShellCheck to GitHub actions ([#240](https://github.com/opendevstack/ods-pipeline/issues/240))
 - Supply default `sonar-project.properties` when none is present, configuring SonarQube out-of-the-box ([#296](https://github.com/opendevstack/ods-pipeline/issues/296))
+- Linting stage in TypeScript build task ([#325](https://github.com/opendevstack/ods-pipeline/issues/325))
 
 ### Changed